+++ src/util.rs
@@ -1,90 +1,124 @@
 use flate2::read::ZlibDecoder;
-use nom::{error::Error, IResult};
+use nom::{
+    error::{ContextError, ErrorKind, ParseError},
+    IResult,
+};
 use std::io::Read;
 use std::path::{Path, PathBuf};
 
-pub fn parse_leb128(mut input: &[u8]) -> IResult<&[u8], u64> {
-    let x = leb128::read::unsigned(&mut input).unwrap();
+pub fn parse_leb128<'a, E>(mut input: &'a [u8]) -> IResult<&'a [u8], u64, E>
+where
+    E: ParseError<&'a [u8]> + ContextError<&'a [u8]>,
+{
+    let start = input;
+    let x = leb128::read::unsigned(&mut input).map_err(|e| {
+        use leb128::read::Error as LebError;
+        let kind = match e {
+            LebError::Overflow => ErrorKind::Satisfy,
+            // Here our Read impl is a slice so only one error possible
+            LebError::IoError(_) => ErrorKind::Eof,
+        };
+        nom::Err::Error(E::from_error_kind(start, kind))
+    })?;
     Ok((input, x))
 }
 
-pub fn parse_string_ref(input: &[u8]) -> IResult<&[u8], String> {
-    let (input, uncompressed_size) = parse_leb128(input)?;
-    let (input, compressed_size) = parse_leb128(input)?;
+pub fn parse_string_ref<'a, E>(input: &'a [u8]) -> IResult<&'a [u8], String, E>
+where
+    E: ParseError<&'a [u8]> + ContextError<&'a [u8]>,
+{
+    let (input, uncompressed_size) = parse_leb128::<E>(input)?;
+    let (input, compressed_size) = parse_leb128::<E>(input)?;
     if compressed_size != 0 {
-        let compressed_size = compressed_size as usize;
-        let mut decoder = ZlibDecoder::new(&input[..compressed_size]);
-        let mut output = vec![];
-        decoder.read_to_end(&mut output).unwrap();
-        let name = String::from_utf8(output);
-        let name = name.unwrap();
-        Ok((&input[compressed_size..], name))
+        if compressed_size as usize > input.len() {
+            Err(nom::Err::Error(E::from_error_kind(input, ErrorKind::Eof)))
+        } else {
+            let compressed_size = compressed_size as usize;
+            let mut decoder = ZlibDecoder::new(&input[..compressed_size]);
+            let mut output = vec![];
+            if let Ok(_) = decoder.read_to_end(&mut output) {
+                let name = String::from_utf8(output);
+                let name = name.unwrap();
+                Ok((&input[compressed_size..], name))
+            } else {
+                let inner = E::from_error_kind(input, ErrorKind::Satisfy);
+                Err(nom::Err::Failure(E::add_context(
+                    input,
+                    "invalid deflate stream",
+                    inner,
+                )))
+            }
+        }
     } else {
         let uncompressed_size = uncompressed_size as usize;
-        let name = String::from_utf8(input[..uncompressed_size].to_vec()).unwrap();
-        Ok((&input[uncompressed_size..], name))
+        match String::from_utf8(input[..uncompressed_size].to_vec()) {
+            Ok(name) => Ok((&input[uncompressed_size..], name)),
+            Err(_e) => {
+                let inner = E::from_error_kind(input, ErrorKind::Satisfy);
+                Err(nom::Err::Error(E::add_context(
+                    input,
+                    "invalid utf-8 string",
+                    inner,
+                )))
+            }
+        }
     }
 }
 
 /// Parses a list of paths - this is currently only used in parsing the sections in an instrumented
 /// object file, and due to CWD joining is different to the other string parsing implemented
-pub fn parse_path_list(input: &[u8], version: u64) -> IResult<&[u8], Vec<PathBuf>> {
-    let (input, list_length) = parse_leb128(input)?;
+pub fn parse_path_list<'a, E>(input: &'a [u8], version: u64) -> IResult<&'a [u8], Vec<PathBuf>, E>
+where
+    E: ParseError<&'a [u8]> + ContextError<&'a [u8]>,
+{
+    let (input, list_length) = parse_leb128::<E>(input)?;
 
     if version < 3 {
         // read_uncompressed
         let (input, values) = parse_uncompressed_file_list(input, list_length, version)?;
         Ok((input, values))
     } else {
-        let (input, uncompressed_size) = parse_leb128(input)?;
-        let (input, compressed_size) = parse_leb128(input)?;
+        let (input, uncompressed_size) = parse_leb128::<E>(input)?;
+        let (input, compressed_size) = parse_leb128::<E>(input)?;
 
         let compressed_size = compressed_size as usize;
         let uncompressed_size = uncompressed_size as usize;
 
         if compressed_size == 0 {
-            let (input, values) = parse_uncompressed_file_list(input, list_length, version)?;
+            let (input, values) = parse_uncompressed_file_list::<E>(input, list_length, version)?;
             Ok((input, values))
         } else {
             let mut decoder = ZlibDecoder::new(&input[..compressed_size]);
             let mut output = vec![];
             output.reserve(uncompressed_size);
             decoder.read_to_end(&mut output).unwrap();
-            let (_compressed_input, values) = match parse_uncompressed_string_list(&output[..]) {
-                Ok((i, v)) => (i, v.iter().map(PathBuf::from).collect()),
-                Err(e) => {
-                    // substitute the decompressed slice start as the error location
-                    let e = match e {
-                        nom::Err::Error(e) => nom::Err::Error(Error {
-                            input,
-                            code: e.code,
-                        }),
-                        nom::Err::Failure(e) => nom::Err::Failure(Error {
-                            input,
-                            code: e.code,
-                        }),
-                        nom::Err::Incomplete(n) => nom::Err::Incomplete(n),
-                    };
-                    return Err(e);
-                }
-            };
+            // Use context error to
+            let values = parse_uncompressed_string_list::<()>(&output)
+                .map(|(_, v)| v.iter().map(PathBuf::from).collect())
+                .map_err(|_| nom::Err::Failure(E::from_error_kind(input, ErrorKind::Fail)))?;
             Ok((&input[compressed_size..], values))
         }
     }
 }
 
-fn read_string(bytes: &[u8]) -> IResult<&[u8], String> {
-    let (bytes, len) = parse_leb128(bytes)?;
+fn read_string<'a, E>(bytes: &'a [u8]) -> IResult<&'a [u8], String, E>
+where
+    E: ParseError<&'a [u8]> + ContextError<&'a [u8]>,
+{
+    let (bytes, len) = parse_leb128::<E>(bytes)?;
     let len = len as usize;
     let string = String::from_utf8_lossy(&bytes[..len]).to_string();
     Ok((&bytes[len..], string))
 }
 
-fn parse_uncompressed_file_list(
-    mut input: &[u8],
+fn parse_uncompressed_file_list<'a, E>(
+    mut input: &'a [u8],
     list_length: u64,
     version: u64,
-) -> IResult<&[u8], Vec<PathBuf>> {
+) -> IResult<&'a [u8], Vec<PathBuf>, E>
+where
+    E: ParseError<&'a [u8]> + ContextError<&'a [u8]>,
+{
     let mut res = vec![];
     if version < 5 {
         for _ in 0..list_length {
@@ -114,10 +148,13 @@ fn parse_uncompressed_file_list(
     }
 }
 
-fn parse_uncompressed_string_list(mut input: &[u8]) -> IResult<&[u8], Vec<String>> {
+fn parse_uncompressed_string_list<'a, E>(mut input: &'a [u8]) -> IResult<&'a [u8], Vec<String>, E>
+where
+    E: ParseError<&'a [u8]> + ContextError<&'a [u8]>,
+{
     let mut res = vec![];
     while !input.is_empty() {
-        let (bytes, len) = parse_leb128(input)?;
+        let (bytes, len) = parse_leb128::<E>(input)?;
         let len = len as usize;
         let string = String::from_utf8_lossy(&bytes[..len]).to_string();
         res.push(string);